% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gff.R
\name{gffSignature}
\alias{gffSignature}
\title{GFF signature text}
\usage{
gffSignature(gff.table)
}
\arguments{
\item{gff.table}{A \code{gff.table} (\code{data.frame}) with genomic features information.}
}
\value{
A vector of texts, one for each row in \code{gff.table}.
}
\description{
Making a signature text from \code{gff.table} data.
}
\details{
For each row in \code{gff.table} a text is created by pasting the data in
the columns \code{Seqid}, \code{Start}, \code{End} and \code{Strand}. This text should be unique
to each entry in the \code{gff.table}.
}
\examples{
# See the example in the Help-file for readGFF.

}
\seealso{
\code{\link{findOrfs}}, \code{\link{gff2fasta}}.
}
\author{
Lars Snipen.
}
