% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microseq-package.R
\name{microseq}
\alias{microseq}
\title{Basic Biological Sequence Analysis}
\usage{
microseq()
}
\description{
A collection of functions for basic analysis of microbial sequence data.
}
\details{
\tabular{ll}{
Package: \tab microseq\cr
Type: \tab Package\cr
Version: \tab 2.1.7\cr
Date: \tab 2025-09-17\cr
License: \tab GPL-2\cr
}
}
\author{
Lars Snipen, Kristian Hovde Liland\cr
Maintainer: Lars Snipen <lars.snipen@nmbu.no>
}
\keyword{package}
