% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orfs.R
\name{orfSignature}
\alias{orfSignature}
\title{Signature for each ORF}
\usage{
orfSignature(orf.table, full = TRUE)
}
\arguments{
\item{orf.table}{A \code{tibble} with ORF information.}

\item{full}{Logical indicating type of signature.}
}
\value{
A text vector with the \code{Signature} for each ORF
}
\description{
Adds a Signature column to an \code{orf.table}.
}
\details{
The \code{orf.table} is a GFF-formatted table with ORF information,
see \code{\link{findOrfs}}.

The \code{Signature} column added by this function contains texts identifying
each ORF in the table.

The full signature (\code{full = TRUE}) contains the \code{Seqid}, \cod{Start},
\code{End} and \code{Strand} information for each ORF, and is always unique 
to each ORF. If \code{full = FALSE} the \code{Signature} will not contain 
the \code{Start} information for each ORF. This means all nested ORFs ending 
at the same stop-codon will get identical \code{Signature}s.
}
\examples{
# Using a genome file in this package
genome.file <- file.path(path.package("microseq"),"extdata","small.fna")

# Reading genome and finding orfs
genome <- readFasta(genome.file)
orf.tbl <- findOrfs(genome)

# Compute signatures
signature.full <- orfSignature(orf.tbl)
signature.reduced <- orfSignature(orf.tbl, full = F)

}
\seealso{
\code{\link{findOrfs}}.
}
\author{
Lars Snipen.
}
