% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta.R
\name{readFasta}
\alias{readFasta}
\alias{writeFasta}
\alias{Fasta}
\title{Read and write FASTA files}
\usage{
readFasta(in.file)
writeFasta(fdta, out.file, width = 80)
}
\arguments{
\item{in.file}{url/directory/name of FASTA file to read.}

\item{fdta}{A \code{\link{tibble}} with sequence data, see \sQuote{Details} below.}

\item{out.file}{Name of FASTA-file to create.}

\item{width}{Number of sequence characters per line.}
}
\value{
\code{\link{readFasta}} returns a \code{\link{tible}} with the contents of the FASTA file stored in
two columns of text. The first, named \samp{Header}, contains
the headerlines and the second, named \samp{Sequence}, contains the sequences.

\code{\link{writeFasta}} produces a FASTA file.
}
\description{
Reads and writes biological sequences (DNA, RNA, protein) in the FASTA format.
}
\details{
These functions handle input/output of sequences in the commonly used FASTA format.
For every sequence it is presumed there is one Header-line starting with a \sQuote{>}.

The sequences are stored in a \code{\link{tibble}}, opening up all the possibilities in R for
fast and easy manipulations. The Header-lines and the sequences are stored in the two columns
named \samp{Header} and \samp{Sequence}, respectively. If other columns are
present, these will be ignored by \code{\link{writeFasta}}.
}
\examples{
\dontrun{
# We need a FASTA-file to read, here is one example file:
ex.file <- file.path(file.path(path.package("microseq"),"extdata"),"small.fasta")
# Reading a file with name in ex.file
fdta <- readFasta(ex.file)
summary(fdta)
}

}
\seealso{
\code{\link{readFastq}}.
}
\author{
Lars Snipen and Kristian Hovde Liland.
}
\keyword{FASTA}
\keyword{Fasta}
\keyword{sequence}
